project: method to return snaps to install
- moved it away from lifecycle
- added support to not install content snaps if a build-snap provides it
